Implemented TC-like creature/gameobjects spawn pools, as well as dynamic respawn time, as well as spawn groups (#6)
* Implemented spawn pools from TC
* Implemented spawn groups system from TC
* Implemented dynamic respawn rate system from TC
* Changed some defaults values for thoses systems
* DB: Set most escort npcs in the "Dynamic Scaling (Escort npc)" group. Those will now trigger respawn time when starting escort.
* DB: Set most collectables nodes in the "Dynamic Scaling (Gathering nodes)" group
* DB: Set most quest npcs in the "Dynamic Scaling (Quest objectives)" group
* Added some info about pools and spawngroups in ".npc info" and ".gobject near" command
* Some more bug fix
* DB: Added some armor for escort npcs with none
* Tweaked a bit TC code for pool respawns

Former-commit-id: 7b515f16710282a797045ad9dd15714c5bdf629f

+++ src/server/game/Storages/WhoListStorage.cpp
@@ -47,7 +47,7 @@ void WhoListStorageMgr::Update()
         }
 
         // Conversion uint32 to uint8 here
-        _whoListStorage.emplace_back(itr->second->GetGUID(), itr->second->GetTeam(), itr->second->GetSession()->GetSecurity(), itr->second->GetLevel(), 
+        _whoListStorage.emplace_back(itr->second->GetGUID(), itr->second->GetTeam(), itr->second->GetSession()->GetSecurity(), uint8(itr->second->GetLevel()), 
             itr->second->GetClass(), itr->second->GetRace(), playerZoneId, itr->second->GetByteValue(PLAYER_BYTES_3, PLAYER_BYTES_3_OFFSET_GENDER), itr->second->IsVisible(),
             widePlayerName, wideGuildName, playerName, guildName);
     }
